there are likely no real changes here, its just stuff I was looking at when I was writing the latest Wiki page talking about Animations.

+++ Animations/Core/StringAnimations.cs
@@ -59,7 +59,7 @@ namespace SmallGalaxy_Engine.Animations
 
         protected override string GetCurrentValue(float progress)
         {
-            return _toAppendTo + _toWrite.Substring(0, (int)((float)_toWrite.Length * progress));
+            return _toAppendTo + _toWrite.Substring(0, (int)(_toWrite.Length * progress));
         }
 
         #endregion // Methods
